auth/requirements.txt
alembic==1.13.2
fastapi==0.115.0
uvicorn[standard]==0.30.0
SQLAlchemy==2.0.32
psycopg2-binary==2.9.11
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
PyJWT==2.9.0
pydantic==2.8.2
pydantic-settings==2.4.0
pytest==9.0.1